Improved error message handling
