+++ datamodel-ui/src/modules/model/model-edit-view.tsx
@@ -5,7 +5,7 @@ import {
   getIsPartOfWithId,
   getOrganizationsWithId,
 } from '@app/common/utils/get-value';
-import getApiError from '@app/common/utils/getApiErrors';
+import getApiError from '@app/common/utils/get-api-errors';
 import {
   translateLanguage,
   translateModelFormErrors,
@@ -168,7 +168,7 @@ export default function ModelEditView({
 
     if (result.isError) {
       const errorMessage = getApiError(result.error);
-      return [...langsWithError, ...otherErrors, errorMessage];
+      return [...langsWithError, ...otherErrors, ...errorMessage];
     }
 
     return [...langsWithError, ...otherErrors];
